% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_functions}
\alias{extract_functions}
\title{Extract functions and roxygen skeletons from R scripts}
\usage{
extract_functions(.paths, .max_token = 2000)
}
\arguments{
\item{.paths}{A character vector of file paths to the R scripts containing the functions.}

\item{.max_token}{An integer specifying the maximum number of tokens (words) per roxygen skeleton.
The default value is 2000.}
}
\value{
A tibble with the following columns:
\describe{
\item{type}{A character vector indicating if the text is a function or dataset text.}
\item{msg_no}{A numeric vector indicating the message number of each roxygen skeleton.}
\item{script}{A character vector indicating the name of the script that the function was extracted from.}
\item{no}{An integer vector indicating the line number of the function within the script.}
\item{name}{A character vector indicating the name of the function.}
\item{text}{A character vector indicating the roxygen skeleton for the function.}
\item{n_token}{An integer vector indicating the number of tokens (words) in the roxygen skeleton.}
\item{n_chars}{An integer vector indicating the number of characters in the roxygen skeleton.}
}
}
\description{
This function takes a list of R scripts and extracts the functions and associated roxygen skeletons.
It returns a tibble with the extracted information for each function, including its message number,
the script name, the line number, the function name, the roxygen skeleton,
the number of tokens (words) in the roxygen skeleton, and the number of characters in the roxygen skeleton.
}
