% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gpt_function}
\alias{gpt_function}
\title{gpt_function}
\usage{
gpt_function(
  .fun,
  .prompt = c("understand", "document", "reference", "roxygen", "unittest", "suggest")
)
}
\arguments{
\item{.fun}{A function to be analyzed. The input function should be a valid R function.}

\item{.prompt}{A character vector specifying the type of question about the input function. The available options are:\cr
\itemize{
\item "understand": Ask clarifying questions to understand the function in its entirety.\cr
\item "document": Write clarifying comments above every single line of the function.\cr
\item "reference": Specify the package of every single function call within the function.\cr
\item "roxygen": Provide a filled-in Roxygen Skeleton (see roxygen2 package).\cr
\item "unittest": Write a unit test for the function.\cr
\item "suggest": Give suggestions on how to improve the function.\cr
}}
}
\value{
A character string containing the generated prompt based on the input function and the selected question type.
}
\description{
This function provides a standardized interface that returns a prompt that can be used for GPT. It takes a function as input and generates a\cr
standardized prompt based on the user's selected question about the function.
}
\examples{
example_function <- function(x) {
  return(x * 2)
}

# Generate a prompt for understanding the example_function
gpt_function(example_function, .prompt = c("understand", "document"))

# Generate a prompt for referencing the example_function
gpt_function(example_function, .prompt = "reference")

# Generate a Roxygen skeleton for the example_function
gpt_function(example_function, .prompt = "roxygen")

# Generate a prompt for writing a unit test for the example_function
gpt_function(example_function, .prompt = "unittest")

# Generate a prompt for suggesting improvements for the example_function
gpt_function(example_function, .prompt = "suggest")
}
