% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gpt_package_prompt}
\alias{gpt_package_prompt}
\title{Generate a GPT-4 prompt for analyzing an R package}
\usage{
gpt_package_prompt(.dir, .max_token = 2000, .output = c("txt", "console"))
}
\arguments{
\item{.dir}{The directory of the R package.}

\item{.max_token}{Maximum tokens per message (integer).}

\item{.output}{Output format: "txt" (text file) or "console" (console output).}
}
\value{
NULL. The function outputs the generated prompt either to a text file or the console, depending on the value of the .output parameter.
}
\description{
This function generates a GPT-4 prompt for analyzing an R package, including its functions and datasets. The prompt can be output to the console or a text file.
}
